+++ spec/unit/installer/user_project_integrator/target_integrator_spec.rb
@@ -338,6 +338,29 @@ module Pod
           )
         end
 
+        it 'adds copy pods resources input and output paths without duplicates' do
+          resource_paths_by_config = {
+            'Debug' => [
+              '${PODS_CONFIGURATION_BUILD_DIR}/DebugLib/SomeBundle.bundle',
+            ],
+            'Release' => [
+              '${PODS_CONFIGURATION_BUILD_DIR}/ReleaseLib/SomeBundle.bundle',
+            ],
+          }
+          @pod_bundle.stubs(:resource_paths_by_config => resource_paths_by_config)
+          @target_integrator.integrate!
+          target = @target_integrator.send(:native_targets).first
+          phase = target.shell_script_build_phases.find { |bp| bp.name == @copy_pods_resources_phase_name }
+          phase.input_paths.sort.should == %w(
+            ${PODS_CONFIGURATION_BUILD_DIR}/DebugLib/SomeBundle.bundle
+            ${PODS_CONFIGURATION_BUILD_DIR}/ReleaseLib/SomeBundle.bundle
+            ${SRCROOT}/../Pods/Target\ Support\ Files/Pods/Pods-resources.sh
+          )
+          phase.output_paths.sort.should == %w(
+            ${TARGET_BUILD_DIR}/${UNLOCALIZED_RESOURCES_FOLDER_PATH}/SomeBundle.bundle
+          )
+        end
+
         it 'does not add embed frameworks build phase input output paths with no frameworks' do
           @pod_bundle.stubs(:framework_paths_by_config => { 'Debug' => {}, 'Release' => {} })
           @target_integrator.integrate!
@@ -388,6 +411,53 @@ module Pod
             ${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/DebugCompiledFramework.framework
             ${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/DebugVendoredFramework.framework
             ${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/ReleaseVendoredFramework.framework
+          )
+        end
+
+        it 'adds embed frameworks build phase input and output paths for vendored and non vendored frameworks without duplicate' do
+          debug_vendored_framework = { :name => 'SomeFramework.framework',
+                                       :input_path => '${PODS_ROOT}/DebugVendoredFramework/ios/SomeFramework.framework',
+                                       :output_path => '${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/SomeFramework.framework',
+                                       :dsym_name => 'SomeFramework.framework.dSYM',
+                                       :dsym_input_path => '${PODS_ROOT}/DebugVendoredFramework/ios/SomeFramework.framework.dSYM',
+                                       :dsym_output_path => '${DWARF_DSYM_FOLDER_PATH}/SomeFramework.framework.dSYM' }
+
+          debug_non_vendored_framework = { :name => 'CompiledFramework.framework',
+                                           :input_path => '${BUILT_PRODUCTS_DIR}/DebugCompiledFramework/CompiledFramework.framework',
+                                           :output_path => '${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/CompiledFramework.framework' }
+
+          release_vendored_framework = { :name => 'ReleaseVendoredFramework.framework',
+                                         :input_path => '${PODS_ROOT}/ReleaseVendoredFramework/ios/SomeFramework.framework',
+                                         :output_path => '${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/SomeFramework.framework',
+                                         :dsym_name => 'SomeFramework.framework.dSYM',
+                                         :dsym_input_path => '${PODS_ROOT}/ReleaseVendoredFramework/ios/SomeFramework.framework.dSYM',
+                                         :dsym_output_path => '${DWARF_DSYM_FOLDER_PATH}/SomeFramework.framework.dSYM' }
+
+          release_non_vendored_framework = { :name => 'CompiledFramework.framework',
+                                             :input_path => '${BUILT_PRODUCTS_DIR}/ReleaseCompiledFramework/CompiledFramework.framework',
+                                             :output_path => '${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/CompiledFramework.framework' }
+
+          framework_paths_by_config = {
+            'Debug' => [debug_vendored_framework, debug_non_vendored_framework],
+            'Release' => [release_vendored_framework, release_non_vendored_framework],
+          }
+          @pod_bundle.stubs(:framework_paths_by_config => framework_paths_by_config)
+          @target_integrator.integrate!
+          target = @target_integrator.send(:native_targets).first
+          phase = target.shell_script_build_phases.find { |bp| bp.name == @embed_framework_phase_name }
+          phase.input_paths.sort.should == %w(
+            ${BUILT_PRODUCTS_DIR}/DebugCompiledFramework/CompiledFramework.framework
+            ${BUILT_PRODUCTS_DIR}/ReleaseCompiledFramework/CompiledFramework.framework
+            ${PODS_ROOT}/DebugVendoredFramework/ios/SomeFramework.framework
+            ${PODS_ROOT}/DebugVendoredFramework/ios/SomeFramework.framework.dSYM
+            ${PODS_ROOT}/ReleaseVendoredFramework/ios/SomeFramework.framework
+            ${PODS_ROOT}/ReleaseVendoredFramework/ios/SomeFramework.framework.dSYM
+            ${SRCROOT}/../Pods/Target\ Support\ Files/Pods/Pods-frameworks.sh
+          )
+          phase.output_paths.sort.should == %w(
+            ${DWARF_DSYM_FOLDER_PATH}/SomeFramework.framework.dSYM
+            ${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/CompiledFramework.framework
+            ${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/SomeFramework.framework
           )
         end